componentize pass - closes #38
